+++ include/string_utils.h
@@ -2,6 +2,8 @@
 #ifndef STRING_UTILS_H
 #define STRING_UTILS_H
 
+/*Includes*/
+#include <stddef.h>
 
 /*Constants*/
 #define OFFSET ('a' - 'A')